+++ client-v3/src/app/tangerine-forms/tangerine-page/tangerine-page.component.html
@@ -1,6 +1,4 @@
 <form class="formly" role="form" novalidate [formGroup]="form" (ngSubmit)="onFormSubmit(model)">
     <formly-form [model]="model" [fields]="config" [form]="form">
-        <button *ngIf="showPreviousButton" type="submit" class="btn btn-default previous">Previous</button>
-        <button type="submit" class="btn btn-default next">Next</button>
     </formly-form>
 </form>
